requirements.txt
Flask
SpeechRecognition
PyAudio
gTTS
pygame
google-genai
python-dotenv
edge-tts
supabase
reportlab
